+++ src/records/services/records.service.ts
@@ -43,10 +43,7 @@ import {
   compareDateAndTime,
   formatNumberToCurrency,
 } from '../../utils';
-import { CreateCategoriesDto } from '../../categories/dtos/categories.dto';
 import { VERSION_RESPONSE } from '../../constants';
-import { SingleCategoryResponse } from '../../categories/interface';
-import { CATEGORY_EXISTS_MESSAGE } from '../../categories/constants';
 import { GeneralResponse } from '../../response.interface';
 
 @Injectable()
@@ -64,14 +61,14 @@ export class RecordsService {
     userId: string,
   ) {
     try {
-      const { category, subCategory, amount } = data;
+      const { category, amount } = data;
       const {
-        data: { category: categoryFoundOrCreated },
-      } = await this.findOrCreateCategoryForRecord(
-        category,
-        subCategory,
+        data: { categories },
+      } = await this.categoriesService.findByNameAndUserId({
+        categoryName: category,
         userId,
-      );
+      });
+      const [categoryFoundOrCreated] = categories;
       const { _id: categoryId } = categoryFoundOrCreated;
       const { fullDate, formattedTime } = formatDateToString(data.date);
       const amountFormatted = formatNumberToCurrency(amount);
@@ -106,12 +103,12 @@ export class RecordsService {
         });
         modelPopulated = await this.incomeModel.populate(modelPopulated, {
           path: 'category',
-          select: '_id categoryName',
+          select: '_id categoryName icon',
         });
       } else {
         modelPopulated = await this.expenseModel.populate(modelSaved, {
           path: 'category',
-          select: '_id categoryName',
+          select: '_id categoryName icon',
         });
       }
 
@@ -124,54 +121,6 @@ export class RecordsService {
         },
         error: null,
       };
-      return response;
-    } catch (error) {
-      throw new BadRequestException(error.message);
-    }
-  }
-
-  /*
-   * This method will take the category only by name.
-   * It will search if the category has been created or not, if it has been created, it will update subcategories if needed.
-   * This method returns the standarized response with the category.
-   */
-  async findOrCreateCategoryForRecord(
-    // category it's only a name
-    category: string,
-    subCategory: string,
-    userId: string,
-  ) {
-    try {
-      // Check if category already exists.
-      const categoryResponse = await this.categoriesService.findByName(
-        category,
-      );
-      const searchedCategory = categoryResponse.data?.categories;
-
-      // The category already exists with that name.
-      if (searchedCategory) {
-        const [foundCategory] = searchedCategory;
-        const response: SingleCategoryResponse = {
-          version: VERSION_RESPONSE,
-          success: true,
-          message: CATEGORY_EXISTS_MESSAGE,
-          data: {
-            category: foundCategory,
-          },
-          error: null,
-        };
-        return response;
-      }
-
-      // The category is a name and does not exists, then create it.
-      const payload: CreateCategoriesDto = {
-        categoryName: category,
-        subCategories: [subCategory],
-      };
-      const response = await this.categoriesService.createOneCategory(
-        payload,
-        userId,
-      );
       return response;
     } catch (error) {
       throw new BadRequestException(error.message);
@@ -252,7 +201,7 @@ export class RecordsService {
         .find({
           account: accountId,
         })
-        .populate('category', 'categoryName')
+        .populate('category', 'categoryName icon')
         .exec();
       const incomes = await this.incomeModel
         .find({
@@ -285,7 +234,7 @@ export class RecordsService {
           account: accountId,
           fullDate: { $regex: new RegExp(regexDate, 'i') },
         })
-        .populate({ path: 'category', select: 'categoryName' })
+        .populate({ path: 'category', select: 'categoryName icon' })
         .exec();
 
       this.verifyExpensesBelongsToUser(expenses, userId);
@@ -321,7 +270,7 @@ export class RecordsService {
           account: accountId,
           fullDate: { $regex: new RegExp(regexDate, 'i') },
         })
-        .populate({ path: 'category', select: 'categoryName' })
+        .populate({ path: 'category', select: 'categoryName icon' })
         .exec();
       const incomes = await this.incomeModel
         .find({
@@ -332,7 +281,7 @@ export class RecordsService {
           path: 'expensesPaid',
           select: '_id shortName amountFormatted fullDate formattedTime',
         })
-        .populate('category', 'categoryName')
+        .populate('category', 'categoryName icon')
         .exec();
 
       this.verifyExpensesBelongsToUser(expenses, userId);
@@ -421,7 +370,6 @@ export class RecordsService {
       const {
         recordId,
         category,
-        subCategory,
         date,
         amount,
         userId: userIdChanges,
@@ -436,14 +384,13 @@ export class RecordsService {
       if (!amount) throw new UnauthorizedException(MISSING_AMOUNT);
 
       const {
-        data: {
-          category: { _id: categoryId, categoryName },
-        },
-      } = await this.findOrCreateCategoryForRecord(
-        category,
-        subCategory,
+        data: { categories },
+      } = await this.categoriesService.findByNameAndUserId({
+        categoryName: category,
         userId,
-      );
+      });
+      const [categoryFoundOrCreated] = categories;
+      const { _id: categoryId, categoryName } = categoryFoundOrCreated;
       const { fullDate, formattedTime } = formatDateToString(date);
       const amountFormatted = formatNumberToCurrency(amount);
       const newChanges = {