+++ CHANGELOG.md
@@ -1,3 +1,33 @@
+## v0.3.2 (2024/03/14)
+
+## Pull Requests
+[#10](https://github.com/RafaelMoro/BE_Personal_Finances/pull/10) |  Fix to make create local categories public
+
+## v0.3.1 (2024/03/12)
+
+### Pull Requests
+[#9](https://github.com/RafaelMoro/BE_Personal_Finances/pull/9) |  Fix local categories getting error if category not found
+
+## v0.3.0 (2024/03/12)
+
+### Pull Requests
+[#8](https://github.com/RafaelMoro/BE_Personal_Finances/pull/8) |  Add service in categories to create local categories.
+
+## v0.2.2 (2024/03/12)
+
+### Pull Requests
+[#7](https://github.com/RafaelMoro/BE_Personal_Finances/pull/7) |  Fix on generating url of forgot password in email
+
+## v0.2.1 (2024/03/09)
+
+### Pull Requests
+[#5](https://github.com/RafaelMoro/BE_Personal_Finances/pull/5) | Fix: Return category icon when fetching records by month and year
+
+## v0.2.0 (2024/03/09)
+
+### Pull Requests
+[#4](https://github.com/RafaelMoro/BE_Personal_Finances/pull/4) | Add icon prop into category entity/DTO
+
 ## v0.1.2 (2024/03/06)
 
 ### Pull Requests